Initial commit. Mostly complete for today.
A set of basic Ruby exercises to get more familiar with functions, and
some basic things with classes.

The exercises are mostly complete as far as today is concerned.

Includes solutions and tests. Will be shifting some files around to make
it easier to run the tests when doing individual exercises.