Add Makefile
Add targets that allow me to create a zip of the exercises for sending.
We copy all the needed files to a new directory, then zip the solutions
file so it isn't immediately accessible (wouldn't want to accidentally
open the solutions file if you didn't mean to), and then zips the whole
thing for easy sending. Finally, we clean up the mess.